Move reset/activation test step definition to Common
It's used by two suites, and clearly related to authentication. It doesn't need its own file.
